Build fixes

+++ lib/ibis/Render/ColorNode.cpp
@@ -250,7 +250,7 @@ namespace ibis
 
         ftk::M44F TintNode::_getColorMatrix() const
         {
-            const float v = _attr->getItem("Value") / 360.F;
+            const float v = static_cast<float>(_attr->getItem("Value")) / 360.F;
             return ftk::tint(v);
         }
